Create an interface for k0s package
