+++ pkg-new/hostutils/interface.go
@@ -3,7 +3,9 @@ package hostutils
 import (
 	"context"
 
+	ecv1beta1 "github.com/replicatedhq/embedded-cluster/kinds/apis/v1beta1"
 	"github.com/replicatedhq/embedded-cluster/pkg/runtimeconfig"
+	"github.com/sirupsen/logrus"
 )
 
 var h HostUtilsInterface
@@ -24,6 +26,7 @@ type HostUtilsInterface interface {
 	ConfigureFirewalld(ctx context.Context, podNetwork, serviceNetwork string) error
 	ResetFirewalld(ctx context.Context) error
 	MaterializeFiles(rc runtimeconfig.RuntimeConfig, airgapBundle string) error
+	CreateSystemdUnitFiles(ctx context.Context, logger logrus.FieldLogger, rc runtimeconfig.RuntimeConfig, isWorker bool, proxy *ecv1beta1.ProxySpec) error
 }
 
 // Convenience functions
@@ -55,4 +58,8 @@ func ResetFirewalld(ctx context.Context) error {
 
 func MaterializeFiles(rc runtimeconfig.RuntimeConfig, airgapBundle string) error {
 	return h.MaterializeFiles(rc, airgapBundle)
+}
+
+func CreateSystemdUnitFiles(ctx context.Context, logger logrus.FieldLogger, rc runtimeconfig.RuntimeConfig, isWorker bool, proxy *ecv1beta1.ProxySpec) error {
+	return h.CreateSystemdUnitFiles(ctx, logger, rc, isWorker, proxy)
 }
