+++ pkg-new/hostutils/system.go
@@ -3,6 +3,7 @@ package hostutils
 import (
 	"bufio"
 	"bytes"
+	"context"
 	_ "embed"
 	"fmt"
 	"os"
@@ -10,8 +11,14 @@ import (
 	"path/filepath"
 	"strconv"
 	"strings"
+	"time"
 
+	"github.com/replicatedhq/embedded-cluster/cmd/installer/goods"
+	ecv1beta1 "github.com/replicatedhq/embedded-cluster/kinds/apis/v1beta1"
 	"github.com/replicatedhq/embedded-cluster/pkg/helpers"
+	"github.com/replicatedhq/embedded-cluster/pkg/helpers/systemd"
+	"github.com/replicatedhq/embedded-cluster/pkg/runtimeconfig"
+	"github.com/sirupsen/logrus"
 	"go.uber.org/multierr"
 )
 
@@ -196,4 +203,144 @@ func ensureKernelModulesLoaded() (finalErr error) {
 func modprobe(module string) error {
 	_, err := helpers.RunCommand("modprobe", module)
 	return err
+}
+
+// CreateSystemdUnitFiles links the k0s systemd unit file. this also creates a new
+// systemd unit file for the local artifact mirror service.
+func (h *HostUtils) CreateSystemdUnitFiles(ctx context.Context, logger logrus.FieldLogger, rc runtimeconfig.RuntimeConfig, isWorker bool, proxy *ecv1beta1.ProxySpec) error {
+	dst := systemdUnitFileName()
+	if _, err := os.Lstat(dst); err == nil {
+		if err := os.Remove(dst); err != nil {
+			return err
+		}
+	}
+	src := "/etc/systemd/system/k0scontroller.service"
+	if isWorker {
+		src = "/etc/systemd/system/k0sworker.service"
+	}
+	if proxy != nil {
+		if err := ensureProxyConfig(fmt.Sprintf("%s.d", src), proxy.HTTPProxy, proxy.HTTPSProxy, proxy.NoProxy); err != nil {
+			return fmt.Errorf("unable to create proxy config: %w", err)
+		}
+	}
+	logger.Debugf("linking %s to %s", src, dst)
+	if err := os.Symlink(src, dst); err != nil {
+		return fmt.Errorf("failed to create symlink: %w", err)
+	}
+
+	if _, err := helpers.RunCommand("systemctl", "daemon-reload"); err != nil {
+		return fmt.Errorf("unable to get reload systemctl daemon: %w", err)
+	}
+	if err := installAndEnableLocalArtifactMirror(ctx, logger, rc); err != nil {
+		return fmt.Errorf("unable to install and enable local artifact mirror: %w", err)
+	}
+	return nil
+}
+
+func systemdUnitFileName() string {
+	return fmt.Sprintf("/etc/systemd/system/%s.service", runtimeconfig.BinaryName())
+}
+
+// ensureProxyConfig creates a new http-proxy.conf configuration file. The file is saved in the
+// systemd directory (/etc/systemd/system/k0scontroller.service.d/).
+func ensureProxyConfig(servicePath string, httpProxy string, httpsProxy string, noProxy string) error {
+	// create the directory
+	if err := os.MkdirAll(servicePath, 0755); err != nil {
+		return fmt.Errorf("unable to create directory: %w", err)
+	}
+
+	// create and write the file
+	content := fmt.Sprintf(`[Service]
+Environment="HTTP_PROXY=%s"
+Environment="HTTPS_PROXY=%s"
+Environment="NO_PROXY=%s"`, httpProxy, httpsProxy, noProxy)
+
+	err := os.WriteFile(filepath.Join(servicePath, "http-proxy.conf"), []byte(content), 0644)
+	if err != nil {
+		return fmt.Errorf("unable to create and write proxy file: %w", err)
+	}
+
+	return nil
+}
+
+// installAndEnableLocalArtifactMirror installs and enables the local artifact mirror. This
+// service is responsible for serving on localhost, through http, all files that are used
+// during a cluster upgrade.
+func installAndEnableLocalArtifactMirror(ctx context.Context, logger logrus.FieldLogger, rc runtimeconfig.RuntimeConfig) error {
+	materializer := goods.NewMaterializer(rc)
+	if err := materializer.LocalArtifactMirrorUnitFile(); err != nil {
+		return fmt.Errorf("failed to materialize artifact mirror unit: %w", err)
+	}
+	if err := writeLocalArtifactMirrorDropInFile(rc); err != nil {
+		return fmt.Errorf("failed to write local artifact mirror environment file: %w", err)
+	}
+	if _, err := helpers.RunCommand("systemctl", "daemon-reload"); err != nil {
+		return fmt.Errorf("unable to get reload systemctl daemon: %w", err)
+	}
+	if _, err := helpers.RunCommand("systemctl", "enable", "local-artifact-mirror"); err != nil {
+		return fmt.Errorf("unable to enable the local artifact mirror service: %w", err)
+	}
+	logger.Debugf("Starting local artifact mirror")
+	if _, err := helpers.RunCommand("systemctl", "start", "local-artifact-mirror"); err != nil {
+		return fmt.Errorf("unable to start the local artifact mirror: %w", err)
+	}
+	if err := waitForLocalArtifactMirror(ctx); err != nil {
+		return fmt.Errorf("unable to wait for the local artifact mirror: %w", err)
+	}
+	logger.Debugf("Local artifact mirror started")
+	return nil
+}
+
+func waitForLocalArtifactMirror(ctx context.Context) error {
+	consecutiveSuccesses := 0
+	requiredSuccesses := 3
+	maxAttempts := 30
+	checkInterval := 2 * time.Second
+
+	var lastErr error
+	for attempt := 0; attempt < maxAttempts; attempt++ {
+		_, err := helpers.RunCommand("systemctl", "status", "local-artifact-mirror")
+		if err == nil {
+			consecutiveSuccesses++
+			if consecutiveSuccesses >= requiredSuccesses {
+				return nil
+			}
+		} else {
+			consecutiveSuccesses = 0
+			lastErr = err
+		}
+
+		select {
+		case <-ctx.Done():
+			return ctx.Err()
+		case <-time.After(checkInterval):
+			continue
+		}
+	}
+
+	return lastErr
+}
+
+const (
+	localArtifactMirrorDropInFileContents = `[Service]
+Environment="LOCAL_ARTIFACT_MIRROR_PORT=%d"
+Environment="LOCAL_ARTIFACT_MIRROR_DATA_DIR=%s"
+# Empty ExecStart= will clear out the previous ExecStart value
+ExecStart=
+ExecStart=%s serve
+`
+)
+
+func writeLocalArtifactMirrorDropInFile(rc runtimeconfig.RuntimeConfig) error {
+	contents := fmt.Sprintf(
+		localArtifactMirrorDropInFileContents,
+		rc.LocalArtifactMirrorPort(),
+		rc.EmbeddedClusterHomeDirectory(),
+		rc.PathToEmbeddedClusterBinary("local-artifact-mirror"),
+	)
+	err := systemd.WriteDropInFile("local-artifact-mirror.service", "embedded-cluster.conf", []byte(contents))
+	if err != nil {
+		return fmt.Errorf("write drop-in file: %w", err)
+	}
+	return nil
 }
