+++ pkg/k0s/k0s.go
@@ -2,39 +2,267 @@ package k0s
 
 import (
 	"context"
+	"encoding/json"
+	"fmt"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"time"
 
 	k0sv1beta1 "github.com/k0sproject/k0s/pkg/apis/k0s/v1beta1"
+	ecv1beta1 "github.com/replicatedhq/embedded-cluster/kinds/apis/v1beta1"
+	"github.com/replicatedhq/embedded-cluster/pkg/airgap"
+	"github.com/replicatedhq/embedded-cluster/pkg/config"
+	"github.com/replicatedhq/embedded-cluster/pkg/helpers"
+	"github.com/replicatedhq/embedded-cluster/pkg/netutils"
+	"github.com/replicatedhq/embedded-cluster/pkg/release"
+	"github.com/replicatedhq/embedded-cluster/pkg/runtimeconfig"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	k8syaml "sigs.k8s.io/yaml"
 )
 
-var (
-	_k0s ClientInterface
+const (
+	k0sBinPath = "/usr/local/bin/k0s"
 )
 
-func init() {
-	Set(&Client{})
+var _ K0sInterface = (*K0s)(nil)
+
+type K0s struct {
 }
 
-func Set(k0s ClientInterface) {
-	_k0s = k0s
+// GetStatus calls the k0s status command and returns information about system init, PID, k0s role,
+// kubeconfig and similar.
+func (k *K0s) GetStatus(ctx context.Context) (*K0sStatus, error) {
+	if _, err := os.Stat(k0sBinPath); err != nil {
+		return nil, fmt.Errorf("%s does not seem to be installed on this node", runtimeconfig.BinaryName())
+	}
+
+	// get k0s status json
+	out, err := exec.CommandContext(ctx, k0sBinPath, "status", "-o", "json").Output()
+	if err != nil {
+		return nil, err
+	}
+	var status K0sStatus
+	err = json.Unmarshal(out, &status)
+	if err != nil {
+		return nil, err
+	}
+	return &status, nil
 }
 
-type K0sStatus struct {
-	Role          string                   `json:"Role"`
-	Vars          K0sVars                  `json:"K0sVars"`
-	ClusterConfig k0sv1beta1.ClusterConfig `json:"ClusterConfig"`
+// Install runs the k0s install command and waits for it to finish. If no configuration
+// is found one is generated.
+func (k *K0s) Install(rc runtimeconfig.RuntimeConfig, networkInterface string) error {
+	ourbin := rc.PathToEmbeddedClusterBinary("k0s")
+	hstbin := runtimeconfig.K0sBinaryPath
+	if err := helpers.MoveFile(ourbin, hstbin); err != nil {
+		return fmt.Errorf("unable to move k0s binary: %w", err)
+	}
+
+	nodeIP, err := netutils.FirstValidAddress(networkInterface)
+	if err != nil {
+		return fmt.Errorf("unable to find first valid address: %w", err)
+	}
+	flags, err := config.InstallFlags(rc, nodeIP)
+	if err != nil {
+		return fmt.Errorf("unable to get install flags: %w", err)
+	}
+	if _, err := helpers.RunCommand(hstbin, flags...); err != nil {
+		return fmt.Errorf("unable to install: %w", err)
+	}
+	if _, err := helpers.RunCommand(hstbin, "start"); err != nil {
+		return fmt.Errorf("unable to start: %w", err)
+	}
+	return nil
 }
 
-type K0sVars struct {
-	AdminKubeConfigPath   string `json:"AdminKubeConfigPath"`
-	KubeletAuthConfigPath string `json:"KubeletAuthConfigPath"`
-	CertRootDir           string `json:"CertRootDir"`
-	EtcdCertDir           string `json:"EtcdCertDir"`
+// IsInstalled checks if the embedded cluster is already installed by looking for
+// the k0s configuration file existence.
+func (k *K0s) IsInstalled() (bool, error) {
+	_, err := os.Stat(runtimeconfig.K0sConfigPath)
+	if err == nil {
+		return true, nil
+	} else if os.IsNotExist(err) {
+		return false, nil
+	}
+
+	return false, fmt.Errorf("unable to check if already installed: %w", err)
 }
 
-type ClientInterface interface {
-	GetStatus(ctx context.Context) (*K0sStatus, error)
+// WriteK0sConfig creates a new k0s.yaml configuration file. The file is saved in the
+// global location (as returned by runtimeconfig.K0sConfigPath). If a file already sits
+// there, this function returns an error.
+func (k *K0s) WriteK0sConfig(ctx context.Context, networkInterface string, airgapBundle string, podCIDR string, serviceCIDR string, overrides string, mutate func(*k0sv1beta1.ClusterConfig) error) (*k0sv1beta1.ClusterConfig, error) {
+	cfgpath := runtimeconfig.K0sConfigPath
+	if _, err := os.Stat(cfgpath); err == nil {
+		return nil, fmt.Errorf("configuration file already exists")
+	}
+	if err := os.MkdirAll(filepath.Dir(cfgpath), 0755); err != nil {
+		return nil, fmt.Errorf("unable to create directory: %w", err)
+	}
+
+	var embCfgSpec *ecv1beta1.ConfigSpec
+	if embCfg := release.GetEmbeddedClusterConfig(); embCfg != nil {
+		embCfgSpec = &embCfg.Spec
+	}
+
+	domains := runtimeconfig.GetDomains(embCfgSpec)
+	cfg := config.RenderK0sConfig(domains.ProxyRegistryDomain)
+
+	address, err := netutils.FirstValidAddress(networkInterface)
+	if err != nil {
+		return nil, fmt.Errorf("unable to find first valid address: %w", err)
+	}
+	cfg.Spec.API.Address = address
+	cfg.Spec.Storage.Etcd.PeerAddress = address
+
+	cfg.Spec.Network.PodCIDR = podCIDR
+	cfg.Spec.Network.ServiceCIDR = serviceCIDR
+
+	if mutate != nil {
+		if err := mutate(cfg); err != nil {
+			return nil, err
+		}
+	}
+
+	cfg, err = applyUnsupportedOverrides(cfg, overrides)
+	if err != nil {
+		return nil, fmt.Errorf("unable to apply unsupported overrides: %w", err)
+	}
+
+	if airgapBundle != "" {
+		// update the k0s config to install with airgap
+		airgap.SetAirgapConfig(cfg)
+	}
+
+	// This is necessary to install the previous version of k0s in e2e tests
+	// TODO: remove this once the previous version is > 1.29
+	unstructured, err := helpers.K0sClusterConfigTo129Compat(cfg)
+	if err != nil {
+		return nil, fmt.Errorf("unable to convert cluster config to 1.29 compat: %w", err)
+	}
+	data, err := k8syaml.Marshal(unstructured)
+	if err != nil {
+		return nil, fmt.Errorf("unable to marshal config: %w", err)
+	}
+	if err := os.WriteFile(cfgpath, data, 0600); err != nil {
+		return nil, fmt.Errorf("unable to write config file: %w", err)
+	}
+	return cfg, nil
 }
 
-func GetStatus(ctx context.Context) (*K0sStatus, error) {
-	return _k0s.GetStatus(ctx)
+// applyUnsupportedOverrides applies overrides to the k0s configuration. Applies the
+// overrides embedded into the binary and then the ones provided by the user (--overrides).
+func applyUnsupportedOverrides(cfg *k0sv1beta1.ClusterConfig, endUserOverridesPath string) (*k0sv1beta1.ClusterConfig, error) {
+	embcfg := release.GetEmbeddedClusterConfig()
+	if embcfg != nil {
+		// Apply vendor k0s overrides
+		vendorOverrides := embcfg.Spec.UnsupportedOverrides.K0s
+		var err error
+		cfg, err = config.PatchK0sConfig(cfg, vendorOverrides, false)
+		if err != nil {
+			return nil, fmt.Errorf("unable to patch k0s config: %w", err)
+		}
+	}
+
+	eucfg, err := helpers.ParseEndUserConfig(endUserOverridesPath)
+	if err != nil {
+		return nil, fmt.Errorf("unable to process overrides file: %w", err)
+	}
+
+	if eucfg != nil {
+		// Apply end user k0s overrides
+		endUserOverrides := eucfg.Spec.UnsupportedOverrides.K0s
+		var err error
+		cfg, err = config.PatchK0sConfig(cfg, endUserOverrides, false)
+		if err != nil {
+			return nil, fmt.Errorf("unable to apply overrides: %w", err)
+		}
+	}
+
+	return cfg, nil
+}
+
+// PatchK0sConfig patches the created k0s config with the unsupported overrides passed in.
+func (k *K0s) PatchK0sConfig(path string, patch string) error {
+	if len(patch) == 0 {
+		return nil
+	}
+	finalcfg := k0sv1beta1.ClusterConfig{
+		ObjectMeta: metav1.ObjectMeta{Name: runtimeconfig.BinaryName()},
+	}
+	if _, err := os.Stat(path); err == nil {
+		data, err := os.ReadFile(path)
+		if err != nil {
+			return fmt.Errorf("unable to read node config: %w", err)
+		}
+		finalcfg = k0sv1beta1.ClusterConfig{}
+		if err := k8syaml.Unmarshal(data, &finalcfg); err != nil {
+			return fmt.Errorf("unable to unmarshal node config: %w", err)
+		}
+	}
+	result, err := config.PatchK0sConfig(finalcfg.DeepCopy(), patch, false)
+	if err != nil {
+		return fmt.Errorf("unable to patch node config: %w", err)
+	}
+	if result.Spec.API != nil {
+		if finalcfg.Spec == nil {
+			finalcfg.Spec = &k0sv1beta1.ClusterSpec{}
+		}
+		finalcfg.Spec.API = result.Spec.API
+	}
+	if result.Spec.Storage != nil {
+		if finalcfg.Spec == nil {
+			finalcfg.Spec = &k0sv1beta1.ClusterSpec{}
+		}
+		finalcfg.Spec.Storage = result.Spec.Storage
+	}
+	if result.Spec.WorkerProfiles != nil {
+		if finalcfg.Spec == nil {
+			finalcfg.Spec = &k0sv1beta1.ClusterSpec{}
+		}
+		finalcfg.Spec.WorkerProfiles = result.Spec.WorkerProfiles
+	}
+	// This is necessary to install the previous version of k0s in e2e tests
+	// TODO: remove this once the previous version is > 1.29
+	unstructured, err := helpers.K0sClusterConfigTo129Compat(&finalcfg)
+	if err != nil {
+		return fmt.Errorf("unable to convert cluster config to 1.29 compat: %w", err)
+	}
+	data, err := k8syaml.Marshal(unstructured)
+	if err != nil {
+		return fmt.Errorf("unable to marshal node config: %w", err)
+	}
+	if err := os.WriteFile(path, data, 0600); err != nil {
+		return fmt.Errorf("unable to write node config file: %w", err)
+	}
+	return nil
+}
+
+// WaitForK0s waits for the k0s API to be available. We wait for the k0s socket to
+// appear in the system and until the k0s status command to finish.
+func (k *K0s) WaitForK0s() error {
+	var success bool
+	for i := 0; i < 30; i++ {
+		time.Sleep(2 * time.Second)
+		spath := runtimeconfig.K0sStatusSocketPath
+		if _, err := os.Stat(spath); err != nil {
+			continue
+		}
+		success = true
+		break
+	}
+	if !success {
+		return fmt.Errorf("timeout waiting for %s", runtimeconfig.BinaryName())
+	}
+
+	for i := 1; ; i++ {
+		_, err := helpers.RunCommand(runtimeconfig.K0sBinaryPath, "status")
+		if err == nil {
+			return nil
+		} else if i == 30 {
+			return fmt.Errorf("unable to get status: %w", err)
+		}
+		time.Sleep(2 * time.Second)
+	}
 }
