+++ pkg/k0s/k0s_test.go
